for #2567, add package of rewrite.builder.sql and rewrite.builder.parameter
